requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
boto3==1.34.0
pyclamd==0.5.0
python-magic==0.4.27
aiofiles==23.2.1
cryptography==41.0.8
python-jose[cryptography]==3.3.0
flower==2.0.1 